created rain.ts and rain service
