+++ app/rain.service.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@angular/core';
+
+import { Rain } from './rain';
+import { RAINS } from './mock-rains';
+
+@Injectable()
+export class RainService {
+  getRains(): Rain[] {
+    return RAINS;
+  }
+}
