+++ npm-debug.log
@@ -1,0 +1,45 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/usr/local/bin/node', '/usr/local/bin/npm', 'start' ]
+2 info using npm@3.10.3
+3 info using node@v6.6.0
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info lifecycle mara@1.0.0~prestart: mara@1.0.0
+6 silly lifecycle mara@1.0.0~prestart: no script for prestart, continuing
+7 info lifecycle mara@1.0.0~start: mara@1.0.0
+8 verbose lifecycle mara@1.0.0~start: unsafe-perm in lifecycle true
+9 verbose lifecycle mara@1.0.0~start: PATH: /usr/local/lib/node_modules/npm/bin/node-gyp-bin:/Users/TLuu/Documents/Sides/mara_angular2/node_modules/.bin:/usr/local/bin:/Library/Frameworks/Python.framework/Versions/3.4/bin:/usr/local/heroku/bin:/Users/TLuu/.rbenv/shims:/Applications/Postgres.app/Contents/Versions/9.5/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/opt/X11/bin:/Library/Frameworks/Python.framework/Versions/3.4/bin:/usr/local/heroku/bin:/Users/TLuu/.rbenv/shims:/Applications/Postgres.app/Contents/Versions/9.5/bin
+10 verbose lifecycle mara@1.0.0~start: CWD: /Users/TLuu/Documents/Sides/mara_angular2
+11 silly lifecycle mara@1.0.0~start: Args: [ '-c', 'tsc && concurrently "npm run tsc:w" "npm run lite" ' ]
+12 silly lifecycle mara@1.0.0~start: Returned: code: 2  signal: null
+13 info lifecycle mara@1.0.0~start: Failed to exec start script
+14 verbose stack Error: mara@1.0.0 start: `tsc && concurrently "npm run tsc:w" "npm run lite" `
+14 verbose stack Exit status 2
+14 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:242:16)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at EventEmitter.emit (events.js:191:7)
+14 verbose stack     at ChildProcess.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/spawn.js:40:14)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at ChildProcess.emit (events.js:191:7)
+14 verbose stack     at maybeClose (internal/child_process.js:877:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
+15 verbose pkgid mara@1.0.0
+16 verbose cwd /Users/TLuu/Documents/Sides/mara_angular2
+17 error Darwin 15.6.0
+18 error argv "/usr/local/bin/node" "/usr/local/bin/npm" "start"
+19 error node v6.6.0
+20 error npm  v3.10.3
+21 error code ELIFECYCLE
+22 error mara@1.0.0 start: `tsc && concurrently "npm run tsc:w" "npm run lite" `
+22 error Exit status 2
+23 error Failed at the mara@1.0.0 start script 'tsc && concurrently "npm run tsc:w" "npm run lite" '.
+23 error Make sure you have the latest version of node.js and npm installed.
+23 error If you do, this is most likely a problem with the mara package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     tsc && concurrently "npm run tsc:w" "npm run lite"
+23 error You can get information on how to open an issue for this project with:
+23 error     npm bugs mara
+23 error Or if that isn't available, you can get their info via:
+23 error     npm owner ls mara
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
